Made the 1-N connection well between wine and winery

+++ WinellyApi/DTOs/Wine/WineDto.cs
@@ -9,7 +9,7 @@ namespace WinellyApi.DTOs.Wine
         public string Type { get; set; }
         public int Year { get; set; }
         public decimal Price { get; set; }
-        public int AlcoholContent { get; set; }
+        public double AlcoholContent { get; set; }
         public int WineryId { get; set; }
     }
 }
